Move the CalendarChart logic to its own handler
