+++ src/App.jsx
@@ -12,6 +12,7 @@ import {
   handleRealizedGainsBySymbolClick, 
   handleRealizedGainsByCategoryClick 
 } from './components/ChartHandler/RealizedGainsHandler.jsx';
+import { handleCalendarChartClick } from './components/ChartHandler/CalendarChartHandler.jsx';
 import './App.css';
 import _ from 'lodash';
 
@@ -21,60 +22,6 @@ function App() {
   const [totals, setTotals] = useState(null);
   const [showChart, setShowChart] = useState(undefined)
   const [chartData, setChartData] = useState([]);
-
-  const formatCalendarChartData = (trades, dividends) => {
-    if (!trades || !trades.length) {
-      return {};
-    }
-
-    // Initialize result object with all months
-    const result = {
-      "1": { value: 0 },
-      "2": { value: 0 },
-      "3": { value: 0 },
-      "4": { value: 0 },
-      "5": { value: 0 },
-      "6": { value: 0 },
-      "7": { value: 0 },
-      "8": { value: 0 },
-      "9": { value: 0 },
-      "10": { value: 0 },
-      "11": { value: 0 },
-      "12": { value: 0  }
-    };
-
-    // Process each trade
-    trades.forEach(trade => {
-      if (trade.datetime && trade.realizedPl) {
-        // Extract month from dateTime (format: "2025-01-03, 11:08:24")
-        const month = parseInt(trade.datetime.split('-')[1]);
-        if (month >= 1 && month <= 12) {
-          result[month].value += trade.realizedPl;
-        }
-      }
-    });
-
-    dividends.forEach(dividend => {
-      const month = parseInt(dividend.date.split('-')[1]);
-      if (month >= 1 && month <= 12) {
-        result[month].value += parseFloat(dividend.amount);
-      }
-    });
-    
-    // Round values to 2 decimal places
-    Object.keys(result).forEach(month => {      
-      result[month].value = new Intl.NumberFormat('en-US', { style: 'currency', currency: 'USD' })
-        .format(Math.round(result[month].value * 100) / 100);
-    });
-
-    return result;
-  }
-
-  const handleCalendarChartClick = () => {
-    const data = formatCalendarChartData(trades, sectionsData.dividends)
-    setChartData(data);
-    setShowChart("calendarChart");
-  }
 
   return (
     <div className="App">
@@ -86,7 +33,7 @@ function App() {
       <button onClick={() => handleRealizedGainsClick(totals, setChartData, setShowChart)}>Realized Gains</button>
       <button onClick={() => handleRealizedGainsBySymbolClick(sectionsData, setChartData, setShowChart)}>Realized Gains by Symbol</button>
       <button onClick={() => handleRealizedGainsByCategoryClick(sectionsData, setChartData, setShowChart)}>Realized Gains by Category / Symbol</button>
-      <button onClick={handleCalendarChartClick}>Calendar Chart</button>
+      <button onClick={() => handleCalendarChartClick(trades, sectionsData, setChartData, setShowChart)}>Calendar Chart</button>
 
       {showChart === "dividendsLineChart" && (
         <DividendLineChart chartData={chartData}/>
